fix NPE when `RequestSerializerFactory` has no ConversionService configured.
