+++ cursorpaging-jpa-api/src/main/java/io/vigier/cursorpaging/jpa/serializer/RequestSerializer.java
@@ -4,7 +4,6 @@ package io.vigier.cursorpaging.jpa.serializer;
 import io.vigier.cursorpaging.jpa.Attribute;
 import io.vigier.cursorpaging.jpa.PageRequest;
 import io.vigier.cursorpaging.jpa.serializer.dto.Cursor;
-import java.time.Instant;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
@@ -15,7 +14,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.SneakyThrows;
 import org.springframework.core.convert.ConversionService;
-import org.springframework.core.convert.TypeDescriptor;
 import org.springframework.lang.Nullable;
 
 @Builder
@@ -31,53 +29,7 @@ public class RequestSerializer<E> {
     private Class<E> entityType;
 
     @Builder.Default
-    private final ConversionService conversionService = new ConversionService() {
-        @Override
-        public boolean canConvert( final Class<?> sourceType, final Class<?> targetType ) {
-            if ( sourceType == null || targetType == null ) {
-                return false;
-            }
-            return sourceType.isAssignableFrom( String.class ) && (targetType.isAssignableFrom( Long.class )
-                    || targetType.isAssignableFrom( Integer.class ) || targetType.isAssignableFrom( Boolean.class )
-                    || targetType.isAssignableFrom( String.class ) || targetType.isAssignableFrom( Object.class ));
-        }
-
-        @Override
-        public boolean canConvert( final TypeDescriptor sourceType, final TypeDescriptor targetType ) {
-            if ( sourceType == null || targetType == null ) {
-                return false;
-            }
-            return canConvert( sourceType.getType(), targetType.getType() );
-        }
-
-        @Override
-        public <T> T convert( final Object source, final Class<T> targetType ) {
-            if ( targetType == String.class ) {
-                return targetType.cast( source.toString() );
-            }
-            if ( targetType == Integer.class ) {
-                return targetType.cast( Integer.valueOf( source.toString() ) );
-            }
-            if ( targetType == Long.class ) {
-                return targetType.cast( Long.valueOf( source.toString() ) );
-            }
-            if ( targetType == Boolean.class ) {
-                return targetType.cast( source.toString().equals( "true" ) );
-            }
-            if ( targetType == Instant.class ) {
-                return targetType.cast( Instant.parse( source.toString() ) );
-            }
-            if ( targetType == Object.class ) {
-                return targetType.cast( source );
-            }
-            return null;
-        }
-
-        @Override
-        public Object convert( final Object source, final TypeDescriptor sourceType, final TypeDescriptor targetType ) {
-            return convert( source, targetType.getObjectType() );
-        }
-    };
+    private final ConversionService conversionService = new SimpleConversionService();
 
     @Builder.Default
     private final Map<String, RuleFactory> filterRuleFactories = new HashMap<>();
